requirements.txt
arrow
garth
pyyaml
requests-oauthlib
stravalib
